requirements.txt
open3d
scikit-learn
imageio
tqdm
numpy
configargparser